requirements.txt
flask
loguru
Flask-WTF
Flask-Bootstrap4
pycodestyle
pycryptodome
pycryptodomex